+++ changelog.md
@@ -6,10 +6,11 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 + Function `get_metric` to define metrics used by `backtesting_forecaster()` and `cv_forecaster()`.
-+ Method `set_out_sample_residuals` now allows to append new residuals.
++ Method `set_out_sample_residuals` now allows to append new residuals, not just overwriting the previous ones.
 
 ### Changed
 
++ `predict` returns a time series instead of numpy array.
 + `backtesting_forecaster` allows to backtest forecasters already trained. Previous implementation of `backtesting_forecaster` always do an initial training. 
 + `backtesting_forecaster` of ForecasterAutoregMultiOutput allows incomplete folds.
 + `backtesting_forecaster` has the option to update out_sample_residuals with backtesting residuals.